requirements.txt
argcomplete==2.0.0
requests==2.28.1
PyYAML==6.0